requirements.txt
asgiref==3.8.1
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
mysqlclient==2.0.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
